Remove StripePrice creation in StripeProduct model
- Modified the StripeProduct model so when it is created it no longer creates the price object. Also removed anything
price related from the StripeProduct model. Did this to keep the StripeProduct and StripePrice more seperated. If
something happens and a StripeProduct is created but there is an error creating the StripePrice for the product,
most of the functionality of the product model would not work as there is no price object associated with it.